add daoadmin sidebar
